+++ src/main/java/com/example/i_commerce/domain/member/service/auth/AuthService.java
@@ -21,13 +21,11 @@ import com.example.i_commerce.domain.member.service.auth.dto.WithDrawRequest;
 import com.example.i_commerce.domain.member.tools.DataEncryptor;
 import com.example.i_commerce.domain.member.tools.EmailHashEncoder;
 import com.example.i_commerce.global.exception.AppException;
-import com.example.i_commerce.global.security.jwt.BlacklistedToken;
 import com.example.i_commerce.global.security.jwt.BlacklistedTokenRepository;
 import com.example.i_commerce.global.security.jwt.JwtTokenUtil;
 import com.example.i_commerce.global.security.jwt.TokenHashUtil;
 import com.example.i_commerce.global.security.jwt.TokenPayload;
 import com.example.i_commerce.global.security.principal.CustomUserPrincipal.PrincipalType;
-import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -125,27 +123,6 @@ public class AuthService {
             email,
             accessToken
         );
-    }
-
-    //로그아웃
-    @Transactional
-    public void logout(String token) {
-        String tokenHash = tokenHashUtil.hash(token);
-        LocalDateTime expiresAt = jwtTokenUtil.getExpiration(token);
-
-        if (blacklistedTokenRepository.existsByTokenHashAndExpiresAtAfter(
-            tokenHash,
-            LocalDateTime.now()
-        )) {
-            return;
-        }
-
-        BlacklistedToken blacklistedToken = BlacklistedToken.builder()
-            .tokenHash(tokenHash)
-            .expiresAt(expiresAt)
-            .build();
-
-        blacklistedTokenRepository.save(blacklistedToken);
     }
 
     //    계정 찾기
