+++ src/main/java/com/example/i_commerce/global/security/jwt/BlacklistedToken.java
@@ -38,6 +38,5 @@ public class BlacklistedToken {
     public BlacklistedToken(String tokenHash, LocalDateTime expiresAt) {
         this.tokenHash = tokenHash;
         this.expiresAt = expiresAt;
-        this.createdAt = LocalDateTime.now();
     }
 }
